MINIFI-554: Move OSUtils to nifi-bootstrap-utils so other assemblies like MiNiFi can use it
