+++ minifi/minifi-assembly/pom.xml
@@ -106,6 +106,10 @@ limitations under the License.
             <type>zip</type>
         </dependency>
         <dependency>
+            <groupId>org.apache.nifi</groupId>
+            <artifactId>nifi-bootstrap-utils</artifactId>
+        </dependency>
+        <dependency>
             <groupId>org.apache.nifi.minifi</groupId>
             <artifactId>minifi-bootstrap</artifactId>
         </dependency>
