+++ minifi/pom.xml
@@ -119,6 +119,11 @@ limitations under the License.
     <dependencyManagement>
         <dependencies>
             <dependency>
+                <groupId>org.apache.nifi</groupId>
+                <artifactId>nifi-bootstrap-utils</artifactId>
+                <version>1.14.0-SNAPSHOT</version>
+            </dependency>
+            <dependency>
                 <groupId>org.apache.nifi.minifi</groupId>
                 <artifactId>minifi-bootstrap</artifactId>
                 <version>1.14.0-SNAPSHOT</version>
