+++ nifi-assembly/pom.xml
@@ -118,6 +118,11 @@ language governing permissions and limitations under the License. -->
         </dependency>
         <dependency>
             <groupId>org.apache.nifi</groupId>
+            <artifactId>nifi-bootstrap-utils</artifactId>
+            <version>1.14.0-SNAPSHOT</version>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.nifi</groupId>
             <artifactId>nifi-resources</artifactId>
             <classifier>resources</classifier>
             <version>1.14.0-SNAPSHOT</version>
@@ -1178,6 +1183,7 @@ language governing permissions and limitations under the License. -->
                                                     <exclude>javax.mail:mail</exclude>
                                                     <!-- must be in lib <exclude>org.apache.nifi:nifi-api</exclude> -->
                                                     <exclude>org.apache.nifi:nifi-bootstrap</exclude>
+                                                    <exclude>org.apache.nifi:nifi-bootstrap-utils</exclude>
                                                     <exclude>org.apache.nifi:nifi-expression-language</exclude>
                                                     <exclude>org.apache.nifi:nifi-parameter</exclude>
                                                     <exclude>org.apache.nifi:nifi-processor-utils</exclude>
@@ -1237,6 +1243,7 @@ language governing permissions and limitations under the License. -->
                                                     <include>javax.mail:mail</include>
                                                     <include>org.apache.nifi:nifi-api</include>
                                                     <include>org.apache.nifi:nifi-bootstrap</include>
+                                                    <include>org.apache.nifi:nifi-bootstrap-utils</include>
                                                     <include>org.apache.nifi:nifi-expression-language</include>
                                                     <include>org.apache.nifi:nifi-parameter</include>
                                                     <include>org.apache.nifi:nifi-processor-utils</include>
