+++ nifi-system-tests/nifi-system-test-suite/pom.xml
@@ -152,6 +152,11 @@
         </dependency>
         <dependency>
             <groupId>org.apache.nifi</groupId>
+            <artifactId>nifi-bootstrap-utils</artifactId>
+            <version>1.14.0-SNAPSHOT</version>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.nifi</groupId>
             <artifactId>nifi-bootstrap</artifactId>
             <version>1.14.0-SNAPSHOT</version>
         </dependency>
